geotiff tests: consolidate reader-path cluster (#2401)
PR 8 of the GeoTIFF test consolidation epic (#2390). Folds 13 old
reader-side test files into eight responsibility-grouped modules under
`xrspatial/geotiff/tests/read/`:

- read/test_basic.py     -- band validation on read.
- read/test_dtypes.py    -- dtype kwarg, float16 promotion, GPU dtype path.
- read/test_compression.py -- codec round-trips and decompression-bomb caps.
- read/test_tiling.py    -- per-tile / per-strip byte caps (CPU + GPU).
- read/test_endianness.py -- big-endian multi-byte GPU read path.
- read/test_nodata.py    -- GPU nodata mask helper.
- read/test_coords.py    -- descending / ascending coord round-trip.
- read/test_streaming.py -- streaming BigTIFF threshold (the one
  cross-directory move from `xrspatial/tests/`).

Tests-only restructure; no source changes. Test bodies are preserved
verbatim except for: dropped issue numbers in names, class grouping
for related cases, fixture filenames stripped of issue numbers, and
relative-import paths adjusted for the new depth (`._helpers.*` ->
`.._helpers.*`).

CLUSTER_AUDIT_PR8.md maps every old `file::test` to its new home and
gets deleted in a follow-up commit before merge.

Closes #2401. References epic #2390. PR 3 (rotated CRS) creates
`read/__init__.py` and `read/test_crs.py` in parallel; whichever PR
merges second resolves the trivial `__init__.py` conflict.
