+++ xrspatial/geotiff/tests/read/test_compression.py
@@ -1,17 +1,13 @@
-"""Tests for decompression-bomb defenses (security finding S1).
+"""Reader compression-codec coverage.
 
-Each codec used by the TIFF reader (deflate, zstd, lz4, packbits) accepts an
-``expected_size`` argument and refuses to produce more than ~5% above that
-size before raising ``ValueError``.  Without these caps a small malicious
-TIFF could expand to many GB during decode and OOM the reader before the
-post-decode size check ran.
+Consolidates:
 
-Each end-to-end test here builds a minimal TIFF that declares a 1024x1024
-uint8 image (1 MiB of legitimate pixel data) and feeds in a strip whose
-decoded size is several MiB. That ratio is enough to trip the cap (~1.05
-MiB) without forcing the test process to allocate a multi-gigabyte
-payload host-side -- the audit's original 1024:1 framing was symbolic;
-what we actually verify is "compressed size << decoded size > cap".
+* ``test_compression.py`` -- codec round-trip unit tests for the
+  reader's decompression entry points (deflate, LZW, predictor encode /
+  decode, the dispatcher).
+* ``test_decompression_caps.py`` -- decompression-bomb defenses
+  (security finding S1) across deflate, ZSTD, LZ4, PackBits, LERC,
+  JPEG 2000, and JPEG.
 """
 from __future__ import annotations
 
@@ -22,8 +18,11 @@ import zlib
 import numpy as np
 import pytest
 
-from xrspatial.geotiff._compression import (deflate_decompress, lz4_decompress, packbits_decompress,
-                                            zstd_decompress)
+from xrspatial.geotiff._compression import (COMPRESSION_DEFLATE, COMPRESSION_LZW, COMPRESSION_NONE,
+                                            compress, decompress, deflate_compress,
+                                            deflate_decompress, lz4_decompress, lzw_compress,
+                                            lzw_decompress, packbits_decompress,
+                                            predictor_decode, predictor_encode, zstd_decompress)
 from xrspatial.geotiff._reader import read_to_array
 
 
@@ -50,8 +49,122 @@ _HAS_PILLOW = _module_available("PIL")
 
 
 # ---------------------------------------------------------------------------
-# Helpers
+# Codec round-trips (formerly test_compression.py)
 # ---------------------------------------------------------------------------
+
+
+class TestDeflate:
+    def test_round_trip(self):
+        data = b'hello world! ' * 100
+        compressed = deflate_compress(data)
+        assert compressed != data
+        assert deflate_decompress(compressed) == data
+
+    def test_empty(self):
+        compressed = deflate_compress(b'')
+        assert deflate_decompress(compressed) == b''
+
+    def test_binary_data(self):
+        data = bytes(range(256)) * 10
+        compressed = deflate_compress(data)
+        assert deflate_decompress(compressed) == data
+
+
+class TestLZW:
+    def test_round_trip_simple(self):
+        data = b'ABCABCABCABC'
+        compressed = lzw_compress(data)
+        decompressed = lzw_decompress(compressed, len(data))
+        assert decompressed.tobytes() == data
+
+    def test_round_trip_repetitive(self):
+        data = b'\x00' * 1000
+        compressed = lzw_compress(data)
+        decompressed = lzw_decompress(compressed, len(data))
+        assert decompressed.tobytes() == data
+
+    def test_round_trip_sequential(self):
+        data = bytes(range(256))
+        compressed = lzw_compress(data)
+        decompressed = lzw_decompress(compressed, len(data))
+        assert decompressed.tobytes() == data
+
+    def test_round_trip_random(self):
+        rng = np.random.RandomState(42)
+        data = bytes(rng.randint(0, 256, size=500, dtype=np.uint8))
+        compressed = lzw_compress(data)
+        decompressed = lzw_decompress(compressed, len(data))
+        assert decompressed.tobytes() == data
+
+    def test_round_trip_large(self):
+        rng = np.random.RandomState(123)
+        data = bytes(rng.randint(0, 256, size=10000, dtype=np.uint8))
+        compressed = lzw_compress(data)
+        decompressed = lzw_decompress(compressed, len(data))
+        assert decompressed.tobytes() == data
+
+    def test_empty(self):
+        compressed = lzw_compress(b'')
+        decompressed = lzw_decompress(compressed, 0)
+        assert decompressed.tobytes() == b''
+
+
+class TestPredictor:
+    def test_round_trip_uint8(self):
+        # 4x4 image, 1 byte per sample
+        data = np.array([10, 20, 30, 40, 50, 60, 70, 80,
+                         90, 100, 110, 120, 130, 140, 150, 160],
+                        dtype=np.uint8)
+        encoded = predictor_encode(data.copy(), 4, 4, 1)
+        decoded = predictor_decode(encoded.copy(), 4, 4, 1)
+        np.testing.assert_array_equal(decoded, data)
+
+    def test_round_trip_float32(self):
+        # 2x3 image, 4 bytes per sample
+        arr = np.array([1.0, 2.0, 3.0, 4.0, 5.0, 6.0], dtype=np.float32)
+        raw = np.frombuffer(arr.tobytes(), dtype=np.uint8).copy()
+        encoded = predictor_encode(raw.copy(), 3, 2, 4)
+        decoded = predictor_decode(encoded.copy(), 3, 2, 4)
+        np.testing.assert_array_equal(decoded, raw)
+
+    def test_predictor_encode_differences(self):
+        # First pixel unchanged, rest are differences
+        data = np.array([10, 20, 30, 40], dtype=np.uint8)
+        encoded = predictor_encode(data.copy(), 4, 1, 1)
+        assert encoded[0] == 10
+        assert encoded[1] == 10  # 20 - 10
+        assert encoded[2] == 10  # 30 - 20
+        assert encoded[3] == 10  # 40 - 30
+
+
+class TestDispatch:
+    def test_none(self):
+        data = b'hello'
+        assert decompress(data, COMPRESSION_NONE).tobytes() == data
+        assert compress(data, COMPRESSION_NONE) == data
+
+    def test_deflate(self):
+        data = b'test data ' * 50
+        compressed = compress(data, COMPRESSION_DEFLATE)
+        assert decompress(compressed, COMPRESSION_DEFLATE).tobytes() == data
+
+    def test_lzw(self):
+        data = b'ABCABC' * 20
+        compressed = compress(data, COMPRESSION_LZW)
+        decompressed = decompress(compressed, COMPRESSION_LZW, len(data))
+        assert decompressed.tobytes() == data
+
+    def test_unsupported(self):
+        with pytest.raises(ValueError, match="Unsupported compression"):
+            decompress(b'', 99)
+        with pytest.raises(ValueError, match="Unsupported compression"):
+            compress(b'', 99)
+
+
+# ---------------------------------------------------------------------------
+# Decompression-bomb caps (formerly test_decompression_caps.py)
+# ---------------------------------------------------------------------------
+
 
 def _build_tiff_with_strip(strip_bytes: bytes, *, compression: int,
                            width: int, height: int) -> bytes:
@@ -64,9 +177,6 @@ def _build_tiff_with_strip(strip_bytes: bytes, *, compression: int,
     """
     bo = '<'
 
-    # Tags: (tag_id, type_id, count, value_or_offset_bytes_4)
-    # We keep every tag's value inline so the TIFF body order is:
-    #   header(8) | IFD | strip
     tags = []
 
     def add_short(tag, val):
@@ -88,11 +198,10 @@ def _build_tiff_with_strip(strip_bytes: bytes, *, compression: int,
 
     tags.sort(key=lambda t: t[0])
     num_entries = len(tags)
-    ifd_size = 2 + 12 * num_entries + 4  # count + entries + next-IFD
+    ifd_size = 2 + 12 * num_entries + 4
     ifd_start = 8
     strip_offset = ifd_start + ifd_size
 
-    # Patch StripOffsets (tag 273) with the real strip location.
     patched = []
     for tag_id, typ, count, raw in tags:
         if tag_id == 273:
@@ -101,22 +210,17 @@ def _build_tiff_with_strip(strip_bytes: bytes, *, compression: int,
     tags = patched
 
     out = bytearray()
-    out += b'II'                                      # little-endian
-    out += struct.pack(f'{bo}H', 42)                  # magic
-    out += struct.pack(f'{bo}I', ifd_start)           # offset to IFD0
-    out += struct.pack(f'{bo}H', num_entries)         # IFD entry count
+    out += b'II'
+    out += struct.pack(f'{bo}H', 42)
+    out += struct.pack(f'{bo}I', ifd_start)
+    out += struct.pack(f'{bo}H', num_entries)
     for tag_id, typ, count, raw in tags:
         out += struct.pack(f'{bo}HHI', tag_id, typ, count)
-        # Pad raw to 4 bytes (all our tags fit inline).
         out += raw.ljust(4, b'\x00')
-    out += struct.pack(f'{bo}I', 0)                    # no next IFD
+    out += struct.pack(f'{bo}I', 0)
     out += strip_bytes
     return bytes(out)
 
-
-# ---------------------------------------------------------------------------
-# Codec-level direct tests
-# ---------------------------------------------------------------------------
 
 # Direct-codec bomb size: 4 MiB of zeros, well above the 1 KiB cap used
 # in those tests but small enough to keep CI host allocations under
@@ -141,14 +245,8 @@ class TestCodecDirect:
         assert out == data
 
     def test_deflate_no_cap_when_expected_size_zero(self):
-        """Backward-compat: ``expected_size=0`` (default) disables the cap.
-
-        Callers that haven't been updated to supply a size must keep
-        getting the unbounded library decode -- otherwise a default
-        cap would silently break them. Round-tripping data larger than
-        any plausible cap proves the disable path is intact.
-        """
-        data = b'A' * (256 * 1024)  # 256 KiB, well above any cap default
+        """Backward-compat: ``expected_size=0`` (default) disables the cap."""
+        data = b'A' * (256 * 1024)
         comp = zlib.compress(data, 9)
         out = deflate_decompress(comp)  # no expected_size kwarg
         assert out == data
@@ -168,8 +266,6 @@ class TestCodecDirect:
         assert out == b'ABCD'
 
     def test_packbits_no_cap_when_expected_size_zero(self):
-        # Same literal-run pattern, no expected_size argument: the
-        # backward-compat path must skip the cap and decode in full.
         data = b'\x03ABCD' * 1024
         out = packbits_decompress(data)
         assert out == b'ABCD' * 1024
@@ -224,15 +320,13 @@ class TestLz4Direct:
 
 
 # ---------------------------------------------------------------------------
-# End-to-end TIFF tests (audit reproducer shape)
+# End-to-end TIFF bomb tests (audit reproducer shape)
 # ---------------------------------------------------------------------------
 
 # 1024 x 1024 uint8 = 1 MiB declared image, cap is ~1.05 MiB. We feed a
 # strip whose decoded size is 8 MiB. The cap is exceeded by ~7x, which is
 # enough to prove the codec rejects rather than silently truncates, while
 # keeping the test's host-side allocation small enough for any CI runner.
-# (The audit's original 1024:1 framing was symbolic; the defense fires the
-# moment decoded > cap, not at any specific ratio.)
 _DECLARED_W = 1024
 _DECLARED_H = 1024
 _DECLARED_BYTES = _DECLARED_W * _DECLARED_H  # 1 MiB
@@ -271,24 +365,16 @@ def test_lz4_bomb_rejected(tmp_path):
     import lz4.frame
     payload = b'\x00' * _BOMB_BYTES
     strip = lz4.frame.compress(payload)
-    # LZ4 has a higher floor than deflate/zstd for runs of zeros, but
-    # still well below the bomb size.
     assert len(strip) < _BOMB_BYTES // 4
     tiff = _build_tiff_with_strip(strip, compression=50004,
                                   width=_DECLARED_W, height=_DECLARED_H)
     path = tmp_path / "lz4_bomb.tif"
     path.write_bytes(tiff)
-    # LZ4 is the Experimental read tier (PR 4 of epic #2340); pass the
-    # opt-in so the test exercises the bomb cap rather than the codec
-    # gate.
     with pytest.raises(ValueError, match="exceed"):
         read_to_array(str(path), allow_experimental_codecs=True)
 
 
 def test_packbits_bomb_rejected(tmp_path):
-    # Packbits "repeat next byte 128 times" header is 0x81 0x00 (2 bytes).
-    # We declare 1024x1024=1 MiB image but supply a 2 MiB strip that
-    # decodes to 128 MiB.  The cap should fire long before allocation.
     strip = b'\x81\x00' * (1024 * 1024)
     tiff = _build_tiff_with_strip(strip, compression=32773,
                                   width=_DECLARED_W, height=_DECLARED_H)
@@ -302,11 +388,11 @@ def test_packbits_bomb_rejected(tmp_path):
 # Negative tests: legitimate high-ratio compression must still pass
 # ---------------------------------------------------------------------------
 
+
 def test_legitimate_high_compression_passes(tmp_path):
     """All-zero array compresses to a fraction of declared size — must pass."""
     arr = np.zeros((_DECLARED_H, _DECLARED_W), dtype=np.uint8)
     strip = zlib.compress(arr.tobytes(), 9)
-    # Confirm we actually have a high ratio (not a degenerate test).
     assert len(strip) < _DECLARED_BYTES // 50
     tiff = _build_tiff_with_strip(strip, compression=8,
                                   width=_DECLARED_W, height=_DECLARED_H)
@@ -319,14 +405,8 @@ def test_legitimate_high_compression_passes(tmp_path):
 
 
 def test_cap_includes_metadata_margin():
-    """The cap allows ~5% of legitimate codec metadata above expected size.
-
-    Some encoders emit small framing or trailing bytes; the cap must not
-    reject them.  We feed a payload exactly at expected_size + a few bytes
-    and confirm it decodes.
-    """
+    """The cap allows ~5% of legitimate codec metadata above expected size."""
     expected = 1000
-    # Decompressed size: expected + 30 (3% over).  Within the 5% margin.
     data = b'A' * (expected + 30)
     comp = zlib.compress(data, 9)
     out = deflate_decompress(comp, expected_size=expected)
@@ -334,31 +414,19 @@ def test_cap_includes_metadata_margin():
 
 
 # ---------------------------------------------------------------------------
-# LERC and JPEG 2000 codec-level bomb tests (issue #1625)
+# LERC and JPEG 2000 codec-level bomb tests
 # ---------------------------------------------------------------------------
-#
-# LERC and JPEG 2000 use external libraries (lerc / glymur) that materialise
-# the full decoded buffer before returning, so the existing post-decode
-# size check in ``_decode_strip_or_tile`` fires only after the bomb is
-# already in memory. The wrappers in ``_compression.py`` now query each
-# codestream's declared dimensions (LERC via ``getLercBlobInfo``, JPEG 2000
-# via ``Jp2k.shape``) and raise before invoking the underlying decoder.
+
 
 @pytest.mark.skipif(not _HAS_LERC, reason="lerc not installed")
 class TestLercDirect:
     def test_lerc_bomb_raises(self):
-        """A LERC blob whose declared dimensions exceed the cap must raise.
-
-        Constant-value rasters compress at >700,000:1 in LERC, so a
-        4096x4096 float32 (64 MiB) encodes to ~94 bytes. The cap is set to
-        1 KiB, well below the declared 64 MiB, and the wrapper must reject
-        the blob before ``lerc.decode`` allocates the output buffer.
-        """
+        """A LERC blob whose declared dimensions exceed the cap must raise."""
         import lerc
         arr = np.zeros((4096, 4096), dtype=np.float32)
         encoded = lerc.encode(arr, 1, False, None, 0.0, 1)
         blob = bytes(encoded[2])
-        assert len(blob) < 1024  # confirm this is a real high-ratio blob
+        assert len(blob) < 1024
         from xrspatial.geotiff._compression import lerc_decompress
         with pytest.raises(ValueError, match="exceed"):
             lerc_decompress(blob, expected_size=1024)
@@ -381,7 +449,7 @@ class TestLercDirect:
         encoded = lerc.encode(arr, 1, False, None, 0.0, 1)
         blob = bytes(encoded[2])
         from xrspatial.geotiff._compression import lerc_decompress
-        out = lerc_decompress(blob)  # no expected_size
+        out = lerc_decompress(blob)
         decoded = np.frombuffer(out, dtype=np.float32).reshape(128, 128)
         assert decoded.shape == arr.shape
 
@@ -389,24 +457,16 @@ class TestLercDirect:
 @pytest.mark.skipif(not _HAS_GLYMUR, reason="glymur not installed")
 class TestJpeg2000Direct:
     def test_jpeg2000_bomb_raises(self, tmp_path):
-        """A JPEG 2000 codestream whose declared shape exceeds the cap raises.
-
-        Glymur reports ``Jp2k(file).shape`` from the SIZ marker without
-        triggering pixel decoding, so the wrapper validates the declared
-        ``H * W * dtype_bytes`` against the bomb cap before calling
-        ``jp2[:]``.
-        """
+        """A JPEG 2000 codestream whose declared shape exceeds the cap raises."""
         import glymur
 
-        # Build a real 2000x2000 uint8 codestream (~150 bytes for zeros).
         arr = np.zeros((2000, 2000), dtype=np.uint8)
         tmp = tmp_path / "src.j2k"
         glymur.Jp2k(str(tmp), data=arr)
         blob = tmp.read_bytes()
-        assert len(blob) < 10_000  # confirm high ratio
+        assert len(blob) < 10_000
         from xrspatial.geotiff._compression import jpeg2000_decompress
         with pytest.raises(ValueError, match="exceed"):
-            # declared output 4 MiB, cap 1 KiB
             jpeg2000_decompress(
                 blob, width=2000, height=2000, samples=1,
                 expected_size=1024)
@@ -415,8 +475,6 @@ class TestJpeg2000Direct:
         """A JPEG 2000 blob whose declared output matches expected_size passes."""
         import glymur
 
-        # Use a 64x64 raster: large enough for the default 6-resolution
-        # OpenJPEG pyramid without tripping its min-tile-size check.
         arr = (np.arange(64 * 64, dtype=np.uint8) % 200).reshape(64, 64)
         tmp = tmp_path / "legit.j2k"
         glymur.Jp2k(str(tmp), data=arr)
@@ -442,14 +500,7 @@ class TestJpeg2000Direct:
 
     def test_jpeg2000_unreadable_shape_fails_closed(
             self, tmp_path, monkeypatch):
-        """If the SIZ marker is unreadable, refuse to call ``jp2[:]``.
-
-        Earlier the wrapper silently disabled the cap on
-        ``Jp2k.shape``/``dtype`` failure, which would let an attacker
-        bypass the bomb guard with a malformed-but-decodable
-        codestream.  The current behaviour is fail-closed: raise
-        ``ValueError`` before any pixel-decoding work runs.
-        """
+        """If the SIZ marker is unreadable, refuse to call ``jp2[:]``."""
         import glymur
         arr = np.zeros((64, 64), dtype=np.uint8)
         tmp = tmp_path / "broken.j2k"
@@ -479,31 +530,15 @@ class TestJpeg2000Direct:
 
 
 # ---------------------------------------------------------------------------
-# JPEG (issue #1792)
+# JPEG SOF cap
 # ---------------------------------------------------------------------------
-#
-# Pillow has its own DecompressionBombError, but it fires only at
-# ~178M pixels (~500 MB RGB). A malicious TIFF can declare a small tile
-# (e.g. 256x256 RGB, ~196 KiB expected) while shipping a JPEG payload
-# whose SOF marker declares a much larger image; that lets ~500 MB
-# allocate per tile before the downstream chunk.size != expected
-# reshape check fires. ``jpeg_decompress`` now parses the JPEG SOF
-# marker and raises before handing the blob to Pillow when the
-# declared output exceeds ``expected * 1.05 + 1`` bytes. See
-# https://github.com/xarray-contrib/xarray-spatial/issues/1792 .
+
 
 def _forge_jpeg_with_sof_dimensions(real_h: int, real_w: int,
                                     real_c: int,
                                     declared_h: int,
                                     declared_w: int) -> bytes:
-    """Build a real JPEG and rewrite the SOF marker's H/W fields.
-
-    Pillow encodes a small valid image so the bytestream is a complete
-    JPEG; we then overwrite the height/width fields in the SOF segment
-    so the SOF claims a much larger image than the payload can decode.
-    The decoder never gets the chance to fail on the mismatch because
-    the pre-decode cap fires first -- which is the property under test.
-    """
+    """Build a real JPEG and rewrite the SOF marker's H/W fields."""
     import io
 
     from PIL import Image
@@ -512,7 +547,6 @@ def _forge_jpeg_with_sof_dimensions(real_h: int, real_w: int,
     buf = io.BytesIO()
     img.save(buf, format='JPEG', quality=75)
     data = bytearray(buf.getvalue())
-    # Find SOF0..SOF3,SOF5..SOF7,SOF9..SOF11,SOF13..SOF15 marker.
     sof_codes = {
         0xC0, 0xC1, 0xC2, 0xC3, 0xC5, 0xC6, 0xC7,
         0xC9, 0xCA, 0xCB, 0xCD, 0xCE, 0xCF,
@@ -524,7 +558,6 @@ def _forge_jpeg_with_sof_dimensions(real_h: int, real_w: int,
             raise AssertionError("forged JPEG lost marker alignment")
         marker = data[i + 1]
         if marker in sof_codes:
-            # SOF: 0xFF Cx | len(2) | precision(1) | H(2) | W(2) | components(1)
             data[i + 5] = (declared_h >> 8) & 0xFF
             data[i + 6] = declared_h & 0xFF
             data[i + 7] = (declared_w >> 8) & 0xFF
@@ -541,24 +574,13 @@ def _forge_jpeg_with_sof_dimensions(real_h: int, real_w: int,
 @pytest.mark.skipif(not _HAS_PILLOW, reason="Pillow not installed")
 class TestJpegDirect:
     def test_jpeg_bomb_raises(self):
-        """A JPEG whose SOF dimensions exceed the per-tile cap must raise.
-
-        The forged JPEG payload itself is small (a real 16x16 image with
-        the SOF marker rewritten to declare 8000x8000x3). The wrapper
-        is given a per-tile expected size of 32x32x3 = 3072 bytes; the
-        declared output 8000*8000*3 = 192_000_000 bytes is well above
-        the ``3072 * 1.05 + 1`` cap and the decode must be refused.
-        """
+        """A JPEG whose SOF dimensions exceed the per-tile cap must raise."""
         blob = _forge_jpeg_with_sof_dimensions(
             real_h=16, real_w=16, real_c=3,
             declared_h=8000, declared_w=8000,
         )
         from xrspatial.geotiff._compression import jpeg_decompress
 
-        # Match the full diagnostic so a regression that swaps in a
-        # different error path (e.g. Pillow's own DecompressionBombError
-        # with a different wording, or a numeric overflow before the
-        # explicit guard) fails the test instead of silently passing.
         with pytest.raises(
             ValueError,
             match=r"jpeg decode would exceed.*Likely a decompression bomb",
@@ -579,20 +601,11 @@ class TestJpegDirect:
         assert arr.shape == (32, 32, 3)
 
     def test_jpeg_no_cap_when_size_kwargs_default(self):
-        """Backward-compat: omitting size kwargs falls back to Pillow's guard.
-
-        Direct callers and round-trip tests pass no dimensions; the
-        pre-check must be a no-op so those keep working.
-        """
+        """Backward-compat: omitting size kwargs falls back to Pillow's guard."""
         blob = _forge_jpeg_with_sof_dimensions(
             real_h=16, real_w=16, real_c=3,
             declared_h=64, declared_w=64,
         )
-        # With no dimension kwargs, the cap is disabled. The forged JPEG
-        # declares 64x64 but encodes only 16x16 of payload -- libjpeg
-        # raises on the truncation; the bomb cap is what we're checking
-        # is *not* the source of any exception here. Catch whatever
-        # Pillow raises and assert it isn't our bomb message.
         from PIL import Image as _Img  # noqa: F401
 
         from xrspatial.geotiff._compression import jpeg_decompress
@@ -601,42 +614,20 @@ class TestJpegDirect:
         except ValueError as exc:
             assert "Likely a decompression bomb" not in str(exc)
         except Exception:
-            # libjpeg/Pillow errors are acceptable -- we only care that
-            # the bomb cap did not fire.
             pass
 
     def test_jpeg_malformed_falls_through_to_pillow(self):
-        """A JPEG without a parseable SOF defers to Pillow's own guard.
-
-        We don't want the pre-check to misclassify weird-but-valid
-        streams; if the helper can't read the SOF it should return
-        ``None`` and let Pillow raise its own error.
-        """
-        # SOI followed by EOI -- a syntactically valid but empty stream
-        # with no SOF marker.
+        """A JPEG without a parseable SOF defers to Pillow's own guard."""
         blob = bytes([0xFF, 0xD8, 0xFF, 0xD9])
         from xrspatial.geotiff._compression import jpeg_decompress
 
-        # No SOF -> bomb cap returns None -> Pillow raises on the empty
-        # stream.
         with pytest.raises(Exception):
             jpeg_decompress(blob, width=32, height=32, samples=3)
 
     def test_jpeg_sof_with_truncated_segment_length_returns_none(self):
-        """A SOF segment whose declared length runs past EOF returns None.
-
-        Without segment-length validation, ``_read_jpeg_sof`` would happily
-        read height/width/components at fixed offsets even when those
-        offsets pointed past the segment. The pre-check now demands
-        ``seg_len >= 8`` and ``i + 2 + seg_len <= n`` before reading;
-        truncated SOFs are treated as "unknown size" and the bomb cap
-        defers to Pillow.
-        """
+        """A SOF segment whose declared length runs past EOF returns None."""
         from xrspatial.geotiff._compression import _read_jpeg_sof
 
-        # SOI | SOF0 | seg_len=64 (advertises 64 bytes of segment, but
-        # the buffer ends after only 10 bytes of segment payload).
-        # Truncation -> _read_jpeg_sof must return None.
         truncated = bytes([
             0xFF, 0xD8,                  # SOI
             0xFF, 0xC0,                  # SOF0
