+++ xrspatial/geotiff/tests/read/test_nodata.py
@@ -1,32 +1,24 @@
-"""Regression tests for issue #1934.
+"""Nodata propagation on read.
 
-``_apply_nodata_mask_gpu`` used to replace the sentinel pixels via
-``cupy.where(arr_gpu == sentinel, nan, arr_gpu)``, which allocates a fresh
-output buffer the same shape as the input. Every call site passes a
-freshly decoded GPU buffer that no caller-visible state aliases, so the
-fix writes NaN into the existing buffer with ``cupy.putmask`` and drops
-one chunk-sized device allocation per call.
+Consolidates the GPU nodata-mask reader coverage:
 
-Two guards here:
-
-1. Correctness -- float and integer paths match the pre-fix behaviour on
-   a representative input (sentinel masked to NaN, non-sentinel pixels
-   preserved, integer dtype promoted to float64).
-2. In-place mutation -- on the float path the output array shares the
-   same device pointer as the input, confirming no fresh allocation. The
-   integer path still allocates via ``astype(float64)``; the test checks
-   the post-astype buffer is then mutated in place rather than copied
-   again by ``cupy.where``.
+* ``test_apply_nodata_mask_gpu_inplace_1934.py`` -- in-place mask
+  semantics for ``_apply_nodata_mask_gpu`` (float and integer paths).
+* ``test_apply_nodata_mask_gpu_with_presence_removed_2208.py`` -- the
+  removed sibling helper stays gone after #2207 wired every GPU eager
+  site through ``_finalize_eager_read``.
 """
 from __future__ import annotations
 
 import numpy as np
+import pytest
 
+from xrspatial.geotiff._backends import _gpu_helpers
 from xrspatial.geotiff.tests.conftest import requires_gpu as _gpu_only
 
 
 @_gpu_only
-def test_apply_nodata_mask_gpu_float_masks_sentinel_to_nan_1934():
+def test_apply_nodata_mask_gpu_float_masks_sentinel_to_nan():
     """Float path masks the sentinel to NaN and leaves other pixels alone."""
     import cupy
 
@@ -45,13 +37,8 @@ def test_apply_nodata_mask_gpu_float_masks_sentinel_to_nan_1934():
 
 
 @_gpu_only
-def test_apply_nodata_mask_gpu_float_in_place_no_copy_1934():
-    """Float path mutates the input buffer in place.
-
-    Before the fix, ``cupy.where`` returned a fresh array, so ``out`` had
-    a different device pointer than ``arr_gpu``. After the fix the input
-    buffer is reused and the pointers match.
-    """
+def test_apply_nodata_mask_gpu_float_in_place_no_copy():
+    """Float path mutates the input buffer in place."""
     import cupy
 
     from xrspatial.geotiff import _apply_nodata_mask_gpu
@@ -65,14 +52,8 @@ def test_apply_nodata_mask_gpu_float_in_place_no_copy_1934():
 
 
 @_gpu_only
-def test_apply_nodata_mask_gpu_float_alloc_count_unchanged_1934():
-    """Float path does not pull a fresh chunk-sized buffer from the pool.
-
-    Uses an isolated ``MemoryPool`` and measures ``total_bytes`` (which
-    counts free blocks too) after a ``free_all_blocks`` so the pre-fix
-    ``cupy.where`` allocation cannot be masked by the input buffer being
-    refcount-freed back to the pool before the assertion.
-    """
+def test_apply_nodata_mask_gpu_float_alloc_count_unchanged():
+    """Float path does not pull a fresh chunk-sized buffer from the pool."""
     import cupy
 
     from xrspatial.geotiff import _apply_nodata_mask_gpu
@@ -81,7 +62,6 @@ def test_apply_nodata_mask_gpu_float_alloc_count_unchanged_1934():
     prev_allocator = cupy.cuda.get_allocator()
     cupy.cuda.set_allocator(isolated_pool.malloc)
     try:
-        # Large enough that an extra allocation would be visible.
         arr_gpu = cupy.full((512, 512), -9999.0, dtype=cupy.float32)
         arr_gpu[0, 0] = 1.0  # plant a non-sentinel pixel
 
@@ -94,17 +74,12 @@ def test_apply_nodata_mask_gpu_float_alloc_count_unchanged_1934():
         isolated_pool.free_all_blocks()
         total_after = isolated_pool.total_bytes()
 
-        # The mask is a transient bool array (1/4 the byte count of float32),
-        # so total_bytes can rise by the mask size but must not rise by the
-        # array's full byte count. Pre-fix would add at least one float32
-        # buffer of the same shape (512*512*4 = 1 MiB).
         array_bytes = arr_gpu.nbytes
         growth = total_after - total_before
         assert growth < array_bytes, (
             f"unexpected allocation growth {growth} bytes >= "
             f"array_bytes {array_bytes}; in-place mutation regressed"
         )
-        # And the returned buffer is the same one we passed in.
         assert out.data.ptr == arr_gpu.data.ptr
     finally:
         cupy.cuda.set_allocator(prev_allocator)
@@ -112,7 +87,7 @@ def test_apply_nodata_mask_gpu_float_alloc_count_unchanged_1934():
 
 
 @_gpu_only
-def test_apply_nodata_mask_gpu_int_promotes_and_masks_1934():
+def test_apply_nodata_mask_gpu_int_promotes_and_masks():
     """Integer path still promotes to float64 and masks the sentinel."""
     import cupy
 
@@ -131,13 +106,8 @@ def test_apply_nodata_mask_gpu_int_promotes_and_masks_1934():
 
 
 @_gpu_only
-def test_apply_nodata_mask_gpu_int_no_extra_buffer_after_astype_1934():
-    """Integer path: only the ``astype(float64)`` buffer is allocated.
-
-    Before the fix the trailing ``cupy.where`` allocated a second
-    chunk-sized float64 buffer. After the fix the ``astype`` buffer is
-    mutated in place.
-    """
+def test_apply_nodata_mask_gpu_int_no_extra_buffer_after_astype():
+    """Integer path: only the ``astype(float64)`` buffer is allocated."""
     import cupy
 
     from xrspatial.geotiff import _apply_nodata_mask_gpu
@@ -158,13 +128,8 @@ def test_apply_nodata_mask_gpu_int_no_extra_buffer_after_astype_1934():
         isolated_pool.free_all_blocks()
         total_after = isolated_pool.total_bytes()
 
-        # Required: one float64 buffer (512*512*8 = 2 MiB) from astype.
-        # Pre-fix would have allocated a second float64 buffer for
-        # cupy.where (another 2 MiB) on top of that.
         float64_bytes = out.nbytes
         growth = total_after - total_before
-        # Allow some slack for the bool mask + .any() scalar (well under
-        # one float64 buffer of slack).
         assert growth < 2 * float64_bytes, (
             f"unexpected allocation growth {growth} bytes >= "
             f"2 * float64_bytes {2 * float64_bytes}; pre-fix double-alloc"
@@ -175,7 +140,7 @@ def test_apply_nodata_mask_gpu_int_no_extra_buffer_after_astype_1934():
 
 
 @_gpu_only
-def test_apply_nodata_mask_gpu_float_nan_sentinel_noop_1934():
+def test_apply_nodata_mask_gpu_float_nan_sentinel_noop():
     """NaN nodata on a float array stays a no-op."""
     import cupy
 
@@ -186,13 +151,12 @@ def test_apply_nodata_mask_gpu_float_nan_sentinel_noop_1934():
     )
     input_ptr = arr_gpu.data.ptr
     out = _apply_nodata_mask_gpu(arr_gpu, float('nan'))
-    # Same buffer back, untouched.
     assert out.data.ptr == input_ptr
     np.testing.assert_array_equal(out.get(), [[1.0, 2.0], [3.0, 4.0]])
 
 
 @_gpu_only
-def test_apply_nodata_mask_gpu_none_nodata_passthrough_1934():
+def test_apply_nodata_mask_gpu_none_nodata_passthrough():
     """``nodata is None`` returns the input array untouched."""
     import cupy
 
@@ -203,3 +167,22 @@ def test_apply_nodata_mask_gpu_none_nodata_passthrough_1934():
     out = _apply_nodata_mask_gpu(arr_gpu, None)
     assert out.data.ptr == input_ptr
     assert out.dtype == cupy.int32
+
+
+# ---------------------------------------------------------------------------
+# Helper removal pin (#2208)
+# ---------------------------------------------------------------------------
+
+
+def test_apply_nodata_mask_gpu_with_presence_not_importable():
+    """The dead sibling helper stays removed after #2207."""
+    # Covers both module-attribute absence and the import-time surface.
+    with pytest.raises(ImportError):
+        from xrspatial.geotiff._backends._gpu_helpers import \
+            _apply_nodata_mask_gpu_with_presence  # noqa: F401
+
+
+def test_apply_nodata_mask_gpu_still_present():
+    """``_apply_nodata_mask_gpu`` is still on the chunked GPU dask path."""
+    assert hasattr(_gpu_helpers, '_apply_nodata_mask_gpu')
+    assert callable(_gpu_helpers._apply_nodata_mask_gpu)
