+++ xrspatial/geotiff/tests/read/test_tiling.py
@@ -1,36 +1,53 @@
-"""Local-file tile/strip byte-count cap (issue #1664).
+"""Tiled-read paths, tile boundaries, byte caps.
 
-Before #1664, ``XRSPATIAL_COG_MAX_TILE_BYTES`` only fired in the HTTP
-fetch path. A crafted local TIFF with a huge ``TileByteCounts`` /
-``StripByteCounts`` could still feed an enormous slice into the
-decompressor, which can balloon into gigabytes of decoded output even
-when the underlying mmap slice is bounded by the file size.
+Consolidates:
 
-These tests fabricate small COGs / strip-TIFFs, rewrite their byte
-counts to oversized values, and check that the cap raises before the
-decoder runs.
+* ``test_local_tile_byte_cap_1664.py`` -- local-file ``TileByteCounts`` /
+  ``StripByteCounts`` cap and the env-driven override (CPU path).
+* ``test_gpu_tile_byte_cap_2026_05_18.py`` -- the matching GPU eager and
+  dask + GPU chunked paths.
 """
 from __future__ import annotations
+
+import importlib.util
 
 import numpy as np
 import pytest
 import xarray as xr
 
 from xrspatial.geotiff import _reader as _reader_mod
-from xrspatial.geotiff import open_geotiff, to_geotiff
+from xrspatial.geotiff import open_geotiff, read_geotiff_gpu, to_geotiff
 
-from ._helpers.tiff_surgery import patch_byte_counts as _patch_byte_counts  # noqa: E402
+from .._helpers.tiff_surgery import patch_byte_counts as _patch_byte_counts
+
+
+def _gpu_available() -> bool:
+    if importlib.util.find_spec("cupy") is None:
+        return False
+    try:
+        import cupy
+        return bool(cupy.cuda.is_available())
+    except Exception:
+        return False
+
+
+_HAS_GPU = _gpu_available()
+_gpu_only = pytest.mark.skipif(
+    not _HAS_GPU, reason="cupy + CUDA required for the GPU read path",
+)
+
 
 # ---------------------------------------------------------------------------
-# Helpers -- patch in-place IFD entries for tile / strip byte counts
+# Helpers
 # ---------------------------------------------------------------------------
 
 
-def _build_forged_tiled_cog(tmp_path, byte_count_value: int) -> str:
+def _build_forged_tiled_cog(tmp_path, byte_count_value: int,
+                            *, basename: str = "forged_tiles") -> str:
     """Write a real tiled COG, patch every TileByteCounts entry, return path."""
     arr = np.arange(64 * 64, dtype=np.float32).reshape(64, 64)
     da = xr.DataArray(arr, dims=['y', 'x'])
-    path = str(tmp_path / "forged_local_tiles_1664.tif")
+    path = str(tmp_path / f"{basename}.tif")
     to_geotiff(da, path, tile_size=32, compression='deflate')
     with open(path, 'rb') as f:
         data = bytearray(f.read())
@@ -44,9 +61,7 @@ def _build_forged_stripped_tif(tmp_path, byte_count_value: int) -> str:
     """Write a strip-organized TIFF, patch every StripByteCounts entry."""
     arr = np.arange(64 * 64, dtype=np.float32).reshape(64, 64)
     da = xr.DataArray(arr, dims=['y', 'x'])
-    path = str(tmp_path / "forged_local_strips_1664.tif")
-    # tiled=False forces strip layout; deflate gets the decompressor on
-    # the hot path so a huge declared size matters.
+    path = str(tmp_path / "forged_strips.tif")
     to_geotiff(da, path, tiled=False, compression='deflate')
     with open(path, 'rb') as f:
         data = bytearray(f.read())
@@ -64,7 +79,6 @@ def _build_forged_stripped_tif(tmp_path, byte_count_value: int) -> str:
 class TestLocalTileByteCap:
     def test_huge_tile_byte_count_rejected(self, tmp_path, monkeypatch):
         """A local tile with a huge TileByteCount raises before decode."""
-        # 100 MB > the 1 MB cap we set below.
         path = _build_forged_tiled_cog(tmp_path, 100 * 1024 * 1024)
         monkeypatch.setenv('XRSPATIAL_COG_MAX_TILE_BYTES', str(1024 * 1024))
 
@@ -78,7 +92,6 @@ class TestLocalTileByteCap:
         with pytest.raises(ValueError) as excinfo:
             open_geotiff(path)
         msg = str(excinfo.value)
-        # The forged value (52,428,800) and the cap (1,024) both appear.
         assert "52,428,800" in msg or "52428800" in msg
         assert "1,024" in msg or "1024" in msg
         assert "denial-of-service" in msg.lower() or "malformed" in msg
@@ -87,7 +100,7 @@ class TestLocalTileByteCap:
         """Legitimate local reads with the default cap still succeed."""
         arr = np.arange(64 * 64, dtype=np.float32).reshape(64, 64)
         da = xr.DataArray(arr, dims=['y', 'x'])
-        path = str(tmp_path / "normal_local_1664.tif")
+        path = str(tmp_path / "normal_local.tif")
         to_geotiff(da, path, tile_size=32, compression='deflate')
 
         result = open_geotiff(path)
@@ -95,25 +108,14 @@ class TestLocalTileByteCap:
 
     def test_env_override_lifts_cap(self, tmp_path, monkeypatch):
         """A user with legitimate large tiles can lift the cap via env."""
-        # 50 MB declared. With cap=64 MB the read succeeds even though
-        # the underlying compressed slice is smaller (mmap truncates at
-        # EOF).
         path = _build_forged_tiled_cog(tmp_path, 50 * 1024 * 1024)
         monkeypatch.setenv(
             'XRSPATIAL_COG_MAX_TILE_BYTES', str(64 * 1024 * 1024))
 
-        # Read may raise inside the decompressor (the truncated mmap
-        # slice is garbage to deflate) but it must NOT raise the cap
-        # error. The thing we care about is that the cap check passes.
         try:
             open_geotiff(path)
         except ValueError as e:
             assert "exceeds the per-tile safety cap" not in str(e)
-
-
-# ---------------------------------------------------------------------------
-# Strip-organized local reads
-# ---------------------------------------------------------------------------
 
 
 class TestLocalStripByteCap:
@@ -141,13 +143,7 @@ class TestLocalStripByteCap:
 
 
 def test_max_tile_bytes_env_negative_falls_back(monkeypatch):
-    """Negative env value falls back to the default, not a 1-byte cap.
-
-    Earlier drafts clamped to ``max(1, val)`` which made a typo
-    (``XRSPATIAL_COG_MAX_TILE_BYTES=-1``) silently reject every tile.
-    The current policy matches ``_http_timeout_from_env``: any non-
-    positive integer is ignored.
-    """
+    """Negative env value falls back to the default, not a 1-byte cap."""
     monkeypatch.setenv('XRSPATIAL_COG_MAX_TILE_BYTES', '-5')
     assert (
         _reader_mod._max_tile_bytes_from_env()
@@ -170,3 +166,73 @@ def test_max_tile_bytes_env_garbage_falls_back(monkeypatch):
         _reader_mod._max_tile_bytes_from_env()
         == _reader_mod.MAX_TILE_BYTES_DEFAULT
     )
+
+
+# ---------------------------------------------------------------------------
+# GPU eager path: per-tile byte cap
+# ---------------------------------------------------------------------------
+
+
+class TestGpuTileByteCap:
+    @_gpu_only
+    def test_huge_tile_byte_count_rejected(self, tmp_path, monkeypatch):
+        """A local tile with a huge TileByteCount raises before GPU decode."""
+        path = _build_forged_tiled_cog(
+            tmp_path, 100 * 1024 * 1024, basename="forged_gpu_tiles")
+        monkeypatch.setenv("XRSPATIAL_COG_MAX_TILE_BYTES", str(1024 * 1024))
+
+        with pytest.raises(ValueError, match="TileByteCount"):
+            read_geotiff_gpu(path)
+
+    @_gpu_only
+    def test_error_message_names_value_and_cap(self, tmp_path, monkeypatch):
+        path = _build_forged_tiled_cog(
+            tmp_path, 50 * 1024 * 1024, basename="forged_gpu_tiles_msg")
+        monkeypatch.setenv("XRSPATIAL_COG_MAX_TILE_BYTES", str(1024))
+
+        with pytest.raises(ValueError) as excinfo:
+            read_geotiff_gpu(path)
+        msg = str(excinfo.value)
+        assert "52,428,800" in msg or "52428800" in msg
+        assert "1,024" in msg or "1024" in msg
+        assert "denial-of-service" in msg.lower() or "malformed" in msg
+
+    @_gpu_only
+    def test_normal_gpu_read_under_default_cap(self, tmp_path):
+        """Legitimate GPU reads with the default cap still succeed."""
+        arr = np.arange(64 * 64, dtype=np.float32).reshape(64, 64)
+        da = xr.DataArray(arr, dims=["y", "x"])
+        path = str(tmp_path / "normal_gpu.tif")
+        to_geotiff(da, path, tile_size=32, compression="deflate")
+
+        result = read_geotiff_gpu(path)
+        np.testing.assert_array_equal(result.data.get(), arr)
+
+    @_gpu_only
+    def test_env_override_lifts_cap(self, tmp_path, monkeypatch):
+        """A user with legitimate large tiles can lift the cap via env."""
+        path = _build_forged_tiled_cog(
+            tmp_path, 50 * 1024 * 1024, basename="forged_gpu_tiles_override")
+        monkeypatch.setenv(
+            "XRSPATIAL_COG_MAX_TILE_BYTES", str(64 * 1024 * 1024))
+
+        try:
+            read_geotiff_gpu(path)
+        except Exception as exc:
+            assert "exceeds the per-tile safety cap" not in str(exc), (
+                "cap loop fired despite the env override lifting the cap"
+            )
+
+
+class TestGpuChunkedTileByteCap:
+    @_gpu_only
+    def test_chunked_huge_tile_byte_count_rejected(
+            self, tmp_path, monkeypatch):
+        """Sibling check on the dask + GPU chunked path."""
+        path = _build_forged_tiled_cog(
+            tmp_path, 100 * 1024 * 1024, basename="forged_gpu_chunked")
+        monkeypatch.setenv(
+            "XRSPATIAL_COG_MAX_TILE_BYTES", str(1024 * 1024))
+
+        with pytest.raises(ValueError, match="TileByteCount"):
+            read_geotiff_gpu(path, chunks=32)
